Simplify login
This simplifies the login page by redirecting from the login provider
straight to the oauth2 identity provider. This works because the only
identity provider we have is OSM.

Note: This might be a bit complicated if a user is logged into OSM with
a different account than they want associated with teams. Usually the
flow would surface something like "previously you've logged in with
account x, do you want to proceed?" but this is an edge case for users
with multiple OSM accounts.